PackageManager::show_warning : read line not char
Reading the first character of input into a char and using that for comparing
can lead to unexpected behaviour when a user enters a multi-char string.
Reading and comparing the whole input will avoid this.

Example:

```cpp
char c;
std::cout << "Enter y for yes, n for no: " << std::flush;
std::cin >> c;

if (c == 'y') {
    // Do actions
} else if (c == 'n') {
    // Do different actions
} else {
    std::cerr << "bad input! Expected y/n" << std::endl;
    exit(1);
```

If the user enters the string "no thankyou" or "you mad????", the `n`
or `y` branches will be matched respectively despite the input being
semmantically invalid and the `else` branch the expected flow of
execution.